@@ -20,7 +20,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.10+10
 
 environment:
-  sdk: ">=2.17.9 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,7 +30,8 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   # cached_network_image: ^3.2.0
-  cashfree_pg: ^2.0.12+32
+  #cashfree_pg: ^2.0.12+32
+  #flutter_cashfree_pg_sdk: ^2.1.1+31
   cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
@@ -38,7 +39,7 @@ dependencies:
   flutter_inappwebview: ^5.7.2+3
   flutter_polyline_points: ^1.0.0
   flutter_spinkit: ^5.1.0
-  flutter_svg: ^1.1.6
+  flutter_svg: ^2.0.9
   fluttertoast: ^8.1.2
   geolocator: ^9.0.2
   get: ^4.6.5
@@ -50,7 +51,7 @@ dependencies:
   lottie: ^2.2.0
   open_file:
     path: external/open_file
-  optimized_cached_image: ^3.0.1
+  #optimized_cached_image:
   package_info_plus: ^4.2.0
   pdf: ^3.8.4
   #razorpay_flutter: ^1.3.4
@@ -63,6 +64,7 @@ dependencies:
   get_storage: ^2.1.1
   shimmer: ^3.0.0
   dio:
+  google_fonts: ^4.0.4
 
 dev_dependencies:
   flutter_lints: ^2.0.0
