@@ -6,11 +6,13 @@ import 'package:get/get.dart';
 import 'package:rentitezy/home/home_controller/home_controller.dart';
 import 'package:rentitezy/theme/custom_theme.dart';
 import 'package:rentitezy/utils/view/rie_widgets.dart';
-import '../../utils/const/app_urls.dart';
-import '../web_view/webview_payment.dart';
-import 'booking_model.dart';
+import '../../../utils/const/app_urls.dart';
+import '../../web_view/webview_payment.dart';
+import '../model/booking_model.dart';
+import '../model/get_single_booking_model.dart';
 
 class MyBookingController extends GetxController {
+  SingleBookingModel? getSingleBooking;
   var myBookingData = <MyBookingModelData>[].obs;
   RxList<Invoices> invoices = <Invoices>[].obs;
 
@@ -53,6 +55,31 @@ class MyBookingController extends GetxController {
         }
         }
       isLoading(false);
+  }
+
+  Future<void> getSingleBookingDetails({required String bookingId}) async {
+    String url = "${AppUrls.getSingleBooking}/?id=$bookingId";
+    isLoading(true);
+    final response = await homeApiController.apiService.getApiCallWithURL(endPoint: url);
+    /*  var response = await http.get(Uri.parse(AppUrls.myBooking), headers: {
+      'auth-token': GetStorage().read(Constants.token).toString()
+    });*/
+    // var responseData = jsonDecode(response.body);
+    //await Future.delayed(const Duration(seconds: 2));
+    String success = response["message"];
+    if (success == 'success') {
+      getSingleBooking = await SingleBookingModel.fromJson(response);
+      update();
+    }
+    if (response['success']) {
+      var singleBookingData = response["data"] ;
+log(" data ${getSingleBooking?.data?.name}");
+      }
+      else{
+        log('No Booking found');
+        RIEWidgets.getToast(message: 'No Booking found', color: CustomTheme.white);
+      }
+    isLoading(false);
   }
 
   Future<void> fetchBookingInvoices({required String bookingID})async{
