@@ -1,26 +1,31 @@
+import 'dart:developer';
+
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get/get_core/src/get_main.dart';
 import 'package:rentitezy/my_bookings/appbar_widget.dart';
-import 'package:rentitezy/my_bookings/booking_model.dart';
 import 'package:rentitezy/utils/const/widgets.dart';
-
 import '../../theme/custom_theme.dart';
 import '../../ticket/view/get_all_ticket.dart';
-import '../../ticket/view/view_ticket_details.dart';
-import '../my_booking_controller.dart';
+import '../controller/my_booking_controller.dart';
 import 'invoice_screen.dart';
 
 class BookingDetailsPage extends StatelessWidget {
-  final MyBookingModelData bookingModelData;
+ // final MyBookingModelData bookingModelData;
   MyBookingController myBookingController = Get.find();
-   BookingDetailsPage({super.key, required this.bookingModelData});
+   BookingDetailsPage({super.key,
+     //required this.bookingModelData
+   });
+  @override
 
   @override
   Widget build(BuildContext context) {
+
+    var data = myBookingController.getSingleBooking?.data;
+    log('$data');
     return Scaffold(
       appBar: appBarBooking(
-          'Booking Details  ( Flat no. ${bookingModelData.propUnit?.flatNo} )',
+          'Booking Details  ( Flat no. ${data?.id} )',
           context,
           false,
           (() {})),
@@ -28,12 +33,12 @@ class BookingDetailsPage extends StatelessWidget {
         width: getScreenWidth,
         height: getScreenHeight,
         padding: EdgeInsets.symmetric(
-            horizontal: getScreenWidth * 0.035,
+            horizontal: getScreenWidth * 0.03,
             vertical: getScreenWidth * 0.05),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            getPropInfoView(context: context),
+            getPropInfoView(context: context,),
             SizedBox(
               height: getScreenHeight*0.02,
             ),
@@ -81,7 +86,7 @@ class BookingDetailsPage extends StatelessWidget {
 
                 ),
                 onPressed: () async {
-                 await  myBookingController.fetchBookingInvoices(bookingID: bookingModelData.id.toString());
+                 await  myBookingController.fetchBookingInvoices(bookingID: '${data?.id.toString()}');
                  Get.to(InvoiceScreen());
                 },
                 child: const Text('Invoice(s)')),
@@ -91,7 +96,7 @@ class BookingDetailsPage extends StatelessWidget {
 
                   ),
                   onPressed: (){
-                    Get.to(const GetAllTickets());
+                    Get.to(const GetAllTickets(),arguments: myBookingController.getSingleBooking?.data?.id.toString());
                   },
                   child: const Text('Create Ticket')),
             ],)
@@ -101,47 +106,61 @@ class BookingDetailsPage extends StatelessWidget {
     );
   }
 
-  Widget getPropInfoView({required BuildContext context}) {
+  Widget getPropInfoView({required BuildContext context,}) {
     return Container(
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           ClipRRect(
             borderRadius: BorderRadius.all(Radius.circular(10)),
-            child: Container(
-              height: getScreenHeight*0.15,
+            child: Card(
+              clipBehavior: Clip.hardEdge,
+              elevation: 2,
+              child: Container(
+                height: getScreenHeight*0.15,
                 width: getScreenWidth*0.25,
-
-                child: Image.network(bookingModelData.propUnit?.listing?.images !=
+                decoration: BoxDecoration(image: DecorationImage(image: NetworkImage(
+                  myBookingController.getSingleBooking?.data?.propListing !=
+                    null &&
+                    myBookingController.getSingleBooking?.data?.propListing?.images!.length != 0
+                    ? myBookingController.getSingleBooking?.data?.propListing?.images![0].url ?? ''
+                    : '',),fit: BoxFit.cover,)),
+                /*child: Image.network(myBookingController.getSingleBooking?.data?.propListing !=
                             null &&
-                        bookingModelData.propUnit?.listing?.images!.length != 0
-                    ? bookingModelData.propUnit?.listing?.images![0].url ?? ''
-                    : '')),
+                    myBookingController.getSingleBooking?.data?.propListing?.images!.length != 0
+                    ? myBookingController.getSingleBooking?.data?.propListing?.images![0].url ?? ''
+                    : '',),*/
+              ),
+            ),
           ),
-          SizedBox(width: getScreenWidth*0.03,),
+          SizedBox(width: getScreenWidth*0.015,),
           Container(
-
             width: getScreenWidth*0.65,
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               mainAxisAlignment: MainAxisAlignment.start,
               children: [
-
                 Row(
                   children: [
                     Icon(Icons.house,size: getScreenWidth*0.05,color: Colors.grey,),
                     SizedBox(width: getScreenWidth*0.01,),
-                    Text('${bookingModelData.propUnit?.listing?.property?.name ?? ''}',style: TextStyle(color: Colors.black,fontSize: 22,fontWeight: FontWeight.w600),),
+                    Text('${myBookingController.getSingleBooking?.data?.propListing?.property?.name ?? ''}',style: TextStyle(color: Colors.black,fontSize: 22,fontWeight: FontWeight.w600),),
                   ],
                 ),
                 SizedBox(height: getScreenHeight*0.005,),
-                Row(
-                  children: [
-                    Icon(Icons.home_max,size: getScreenWidth*0.05,color: Colors.grey,),
-                    SizedBox(width: getScreenWidth*0.01,),
-                    Text('Flat no. ',style: TextStyle(color: Colors.black54,fontSize: 18,fontWeight: FontWeight.w600),),
-                    Text(bookingModelData.propUnit?.flatNo ?? '',style: TextStyle(color: Colors.black54,fontSize: 18,fontWeight: FontWeight.w600),),
-                  ],
+
+                   Row(
+                     crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Icon(Icons.home_max,size: getScreenWidth*0.05,color: Colors.grey,),
+                      SizedBox(width: getScreenWidth*0.01,),
+                   Container(
+                        width: Get.width*0.52,
+                        child: Text('Address: ${myBookingController.getSingleBooking?.data?.propListing?.property?.address ?? ''}',
+                          maxLines:3,style: TextStyle(color: Colors.black54,fontSize: 16,fontWeight: FontWeight.w600),),
+                      ),
+                    ],
+
                 ),
                 SizedBox(height: getScreenHeight*0.005,),
                 Row(
@@ -152,7 +171,7 @@ class BookingDetailsPage extends StatelessWidget {
                     Container(
 
                         width: getScreenWidth*0.58,
-                        child: Text('${bookingModelData.propUnit?.listing?.property?.address ?? ''}',style: TextStyle(color: Colors.black54,fontSize: 16,fontWeight: FontWeight.w500),)),
+                        child: Text('${myBookingController.getSingleBooking?.data?.propListing?.property?.latlng ?? ''}',style: TextStyle(color: Colors.black54,fontSize: 16,fontWeight: FontWeight.w500),)),
                   ],
                 ),
 
@@ -175,7 +194,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Rent',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.18,),
-              Text('₹ ${bookingModelData.rent ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('₹ ${myBookingController.getSingleBooking?.data?.rent ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -183,7 +202,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Deposit',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.13,),
-              Text('₹ ${bookingModelData.deposit ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('₹ ${myBookingController.getSingleBooking?.data?.deposit ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -191,7 +210,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Onboarding',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.07,),
-              Text('₹ ${bookingModelData.onboarding ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('₹ ${myBookingController.getSingleBooking?.data?.onboarding ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -199,7 +218,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Amount Paid',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.05,),
-              Text('₹ ${bookingModelData.amountPaid ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('₹ ${myBookingController.getSingleBooking?.data?.amountPaid ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
 
@@ -221,7 +240,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Name',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.12,),
-              Text('${bookingModelData.name ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('${myBookingController.getSingleBooking?.data?.name ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -229,7 +248,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Email',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.12,),
-              Text('${bookingModelData.email ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('${myBookingController.getSingleBooking?.data?.email ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -237,7 +256,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Mobile',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.1,),
-              Text('${bookingModelData.phone ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('${myBookingController.getSingleBooking?.data?.phone ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           )
 
@@ -259,7 +278,7 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('No. of Guest',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.07,),
-              Text('${bookingModelData.guest ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('${myBookingController.getSingleBooking?.data?.guest ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -267,7 +286,8 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Move-In',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.14,),
-              Text('${bookingModelData.moveIn ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('${myBookingController.getSingleBooking?.data?.from.toString().split('T')[0]}'
+                ,style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
           SizedBox(height: getScreenHeight*0.005,),
@@ -275,11 +295,9 @@ class BookingDetailsPage extends StatelessWidget {
             children: [
               Text('Move-Out',style: TextStyle(color: Colors.grey,fontSize: 16,fontWeight: FontWeight.w600),),
               SizedBox(width: getScreenWidth*0.115,),
-              Text('${bookingModelData.moveOut ?? ''}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
+              Text('${myBookingController.getSingleBooking?.data?.till.toString().split('T')[0]}',style: TextStyle(color: Colors.black,fontSize: 16,fontWeight: FontWeight.w600),),
             ],
           ),
-
-
         ],
       ),
     );
