@@ -17,8 +17,8 @@ import 'package:rentitezy/screen/thankyou_page.dart';
 import 'package:rentitezy/widgets/const_widget.dart';
 import 'package:http/http.dart' as http;
 import 'package:url_launcher/url_launcher_string.dart';
-import '../my_bookings/booking_model.dart';
-import '../my_bookings/my_booking_controller.dart';
+import '../my_bookings/model/booking_model.dart';
+import '../my_bookings/controller/my_booking_controller.dart';
 import '../utils/const/app_urls.dart';
 import '../utils/const/widgets.dart';
 import '../utils/view/rie_widgets.dart';
